// supabase/functions/send-contact-email/types.ts
export interface ContactEmailRequest {
  nombre: string;
  email: string;
  telefono?: string;
  servicio: string;
  mensaje: string;
  institucion?: string;
  area?: string;
  ubicacion?: string;
  proyecto_nombre?: string;
  proyecto_objetivo?: string;
  proyecto_estado?: string;
  proyecto_fechainicio?: string;
  proyecto_descripcion?: string;
  tipo_impacto?: string;
  grupos_beneficiados?: string;
  estimacion_beneficiarios?: string;
  ubicacion_impacto?: string;
  aliados_colaboradores?: string;
  recursos_adicionales?: string;
  plan_seguimiento?: string;
  plan_sustentabilidad?: string;
  web_o_redes?: string;
  motivacion?: string;
}